+++ common/src/main/java/tocraft/remorphed/command/RemorphedCommand.java
@@ -22,7 +22,7 @@ import net.minecraft.world.entity.LivingEntity;
 import org.jetbrains.annotations.Nullable;
 import tocraft.craftedcore.event.common.CommandEvents;
 import tocraft.remorphed.Remorphed;
-import tocraft.remorphed.impl.RemorphedPlayerDataProvider;
+import tocraft.remorphed.impl.PlayerMorph;
 import tocraft.walkers.api.PlayerShapeChanger;
 import tocraft.walkers.api.variant.ShapeType;
 
@@ -31,7 +31,7 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
         ShapeType<LivingEntity> type = getType(source.getLevel(), id, nbt);
         Component name = new TranslatableComponent(type.getEntityType().getDescriptionId());
 
-        if (((RemorphedPlayerDataProvider) player).remorphed$getUnlockedShapes().containsKey(type)) {
+        if (PlayerMorph.getUnlockedShapes(player).containsKey(type)) {
             source.sendSuccess(new TranslatableComponent(Remorphed.MODID + ".hasShape_success",
                     player.getDisplayName(), name), true);
 
@@ -46,7 +46,7 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
         ShapeType<LivingEntity> type = getType(source.getLevel(), id, nbt);
         Component name = new TranslatableComponent(type.getEntityType().getDescriptionId());
 
-        ((RemorphedPlayerDataProvider) player).remorphed$getUnlockedShapes().remove(type);
+        PlayerMorph.getUnlockedShapes(player).remove(type);
 
         source.sendSuccess(new TranslatableComponent(Remorphed.MODID + ".removeShape", name, player.getDisplayName()), true);
     }
@@ -55,13 +55,13 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
         ShapeType<LivingEntity> type = getType(source.getLevel(), id, nbt);
         Component name = new TranslatableComponent(type.getEntityType().getDescriptionId());
 
-        ((RemorphedPlayerDataProvider) player).remorphed$getUnlockedShapes().put(type, Remorphed.getKillToUnlock(type.getEntityType()));
+        PlayerMorph.getUnlockedShapes(player).put(type, Remorphed.getKillToUnlock(type.getEntityType()));
 
         source.sendSuccess(new TranslatableComponent(Remorphed.MODID + ".addShape", player.getDisplayName(), name), true);
     }
 
     private static void clearShapes(CommandSourceStack source, ServerPlayer player) {
-        ((RemorphedPlayerDataProvider) player).remorphed$getUnlockedShapes().clear();
+        PlayerMorph.getUnlockedShapes(player).clear();
 
         source.sendSuccess(new TranslatableComponent(Remorphed.MODID + ".clearShapes", player.getDisplayName()), true);
         PlayerShapeChanger.change2ndShape(player, null);